Add windscreen to FlightDeck 3D model
To properly apply ALS glass shaders, it appears necessary to have two
objects, one to render the shader on on the exterior view, and another
for the interior view.

To accomodate this, the flightdeck got a windscreen.

Some other things may have slightly changed because I exported the
AC3D file using newer version of the blender exporter.