Rewrite autobrake.nas according to manual
To implement the requirements of the manual, I made some pretty
drastic rewriting of the nasal subsystem.

It's already been tested by yours truly and it works according to the
manual.

This rewriting also created a new position for the autobrake knob, the
DISARM position, it will be animated shortly.