Modify ground effects
The previous ground effects factor wasn't realistic and was causing
the plane to "float" too high on final approach. Ground effects do
cause the plane to "float", but just before touching the ground, not
100ft AGL.

Also, the change in induced drag due to ground effects was not modeled
into the FDM, so it was added.

I referenced these pages for this change:
1. https://mcleodsean.wordpress.com/2018/02/06/landing-ground-effect-flare-jsbsim/
2. https://jsbsim-team.github.io/jsbsim-reference-manual/mypages/user-manual-math/