Use materials' specific heat capacity & more objects
It is possible to simulate more realistic heating and cooling
mechanics by using the heat energy of each of the objects that can be
heated. The heat energy is calculated as:

part's temperture in °K * part's mass * material's specific heat capacity

The FCOM specifies 4 things that can be heated with the anti-ice
system:
1. The leading edge of the wings
2. The engines' nose cowl
3. The windscreen
4. The pitot & AoA probes

Now each of them is being simulated using heat energy.

Also, the simulation of the windscreen temperature will be made in
anti-ice.nas and not heating.nas as it was before, this is because the
FCOM groups engine & wing anti-ice and windscreen anti-ice under the
same anti-ice system.