Use surface area to control parts' cooling
Some parts were cooling too slowly and some too fast.

I tracked down the issue and found that the parts' surface area
exposed to the wind was not taken into account when calculating the
effect of TAT on the part's temperature.

This also meant that the size of the heaters on each part can now be
much more reasonable and in accordance to the part's size.