Animate HDG-TRK selector
This allows pilots to use the mode control panel to switch between
heading and track modes.

it-autoflight will set the heading from 000 to 360 when you switch
between modes while your heading or track is 000. To make this
invisible, the knob was adjusted to set headings from 001 to 360,
instead of 000 to 359.