Improve EGPWS with 2500 call and gear altitude
The 2500 ft AGL call is emitted by real 787s' EGPWS, but this wasn't
configured in the simulation.

We can also take advantage of the gear-agl-ft system provided by
IT-AUTOFLIGHT to get callouts based on the altitude of the gear and
not the CG which is the default.

Additionally, the EGPWS sounds are provided by FlightGear since
2010-07-30, so we don't need to keep a copy.