Better tiller steering internals
The old steer-pos-deg calculations led to a system where the rudder
could maintain some control of the nose wheel even if the slider was
being used to control the tiller or the pushback truck was connected
(and the bypass pin removed).

This was obviously, not good. So new internals were build to fix this,
in this new system each source has a priority, and they override each
other:

1. Pushback truck: It overrides all other steering commands.
2. Tiller control: It overrides the rudder but not the pushback truck.
3. Rudder control: It can only steer 7 degrees, and it has the least
   priority.