Introduce new icing based on icing conditions
Instead of just controlling the "icing" directly based on a part's
temperature. We can only create icing under more "real life" like
conditions, aka temperature below freezing and visible moisture.

This also required changing the specs of the heaters to better suit
the situations under which we can encounter icing.

And I also found that I was making an erroneous calculation of TAT, so
it was fixed.