Update icing warnings in instruments and text
Screen text alerts were still based on temperature, they're now based
on actual icing.

The instrumentation still relied on temperature to report icing. This
is now fixed. I also fixed a bug that could cause eng2 icing not to be
reported, just a reference to an incorrect object in the .aac file.