Limit R-V/S, compensate SBP with throttle, bugfixes
VNAV doesn't lock the autothrottle at max or min when using it's SBP
controller. But the SBP controller was based on ITAF's FLCH, which
works under that assumption. So the controller was updated to take
throttle "vibrations" into account.

The switching between SBP and R-V/S is, for the first time, controlled
enough that I decided to eliminate the final lags to make VNAV more
responsive.

In some situations VNAV can decide to go from SBP to R-V/S, but the
difference between them is so big, that it induces VNAV to switch back
to SBP, and repeat.

To help VNAV find a balance quickly, we can limit the difference
between R-V/S and SBP by limiting R-V/S to +/- 250 fpm relative to
SBP.

Also, there are many bug fixes.