Use new code for VNAV autogen
Through the EFB, pilots can access a function to automatically
generate a VNAV path, while the altitudes generated by this function
were ok, it required pilots to input a cruise altitude, TOC, and TOD
wapoints.

This information was given to the FMC through the CDU, but the VNAV
CONFIG page was overhauled.

So, the FMC got new code capable of automatically determining cruising
altitude, TOC and TOD. It can even handle super short routes and
altitude differences between departure and destination airports.

It also provides smoother vertical paths in climb and descent.

I also found a bug in the EFB's VNAV page while working on this, it
could cause the EFB to crash if pilots edited the route while in the
VNAV page. It was fixed.