Improve handling of climb to descent transition
When the plane fell below the planed climb profile, and the route
transitions from a climb to a descent without cruising, the plane
would descend without authorization, as V/S mode remains engaged from
the previous climb.

To fix this, the R-V/S driver in the transition controller got a new
"block-descent" property, which limits R-V/S such that it can't go
below 0 fpm when in an unauthorized descent. This means the plane will
just maintain altitude until the descent is authorized.